A workaround when kiconthemes is installed
If kiconthemes ≥ 5.80.0 is installed, stylesheet SVG icons will be controlled by it but it has a bug that makes the main menu icon disappear in one LXQt theme. Therefore, a workaround is added to that theme.

NOTE: The bug(s) of kiconthemes ≥ 5.80.0 may also make some icons ugly. Unfortunately, we can do nothing about it.